core/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "erdpy_core"
version = "0.3.0"
authors = [
  { name="MultiversX" },
]
description = "erdpy_core"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
  "pycryptodomex==3.16.0"
]

[project.urls]
"Homepage" = "https://github.com/ElrondNetwork/sdk-erdpy-core"
